LayoutTests/platform/wpe/transforms/2d/transform-value-types-expected.txt

PASS jsWrapperClass(transformRule) is 'CSSValueList'
PASS jsWrapperClass(transformRule.__proto__) is 'CSSValueListPrototype'
PASS jsWrapperClass(transformRule.constructor) is 'Function'
PASS jsWrapperClass(transformRule[0]) is 'CSSValue'
PASS jsWrapperClass(transformRule[0].__proto__) is 'CSSValuePrototype'
PASS jsWrapperClass(transformRule[0].constructor) is 'Function'
PASS transformRule[0].cssText is 'translate(10px)'
PASS transformRule[1].cssText is 'translateX(10px)'
PASS transformRule[2].cssText is 'translateY(10px)'
PASS transformRule[3].cssText is 'rotate(10deg)'
PASS transformRule[4].cssText is 'rotate(11rad)'
PASS transformRule[5].cssText is 'rotate(12grad)'
PASS transformRule[6].cssText is 'rotate(13turn)'
PASS transformRule[7].cssText is 'scale(2)'
PASS transformRule[8].cssText is 'scaleX(1.2)'
PASS transformRule[9].cssText is 'scaleY(1.2)'
PASS transformRule[10].cssText is 'skew(10deg, 10rad)'
PASS transformRule[11].cssText is 'skewX(10deg)'
PASS transformRule[12].cssText is 'skewY(11grad)'
PASS transformRule[13].cssText is 'matrix(1, 0, 0, 1, 0, 0)'
PASS successfullyParsed is true

TEST COMPLETE
